+++ package.json
@@ -12,6 +12,7 @@
     "@googlemaps/google-maps-services-js": "^3.4.2",
     "@prisma/client": "^6.18.0",
     "next": "15.5.4",
+    "next-auth": "^4.24.13",
     "pretendard": "^1.3.9",
     "react": "19.1.0",
     "react-dom": "19.1.0"
