+++ src/app/layout.js
@@ -1,20 +1,19 @@
-// src/app/layout.js
-import "./globals.css";
+import './globals.css';
+import Providers from './providers';
 
 export const metadata = {
-  title: "2099-Trip",
-  description: "Travel helper",
+	title: '2099-Trip',
+	description: 'Travel helper'
 };
 
 export default function RootLayout({ children }) {
-  return (
-    <html lang="ko">
-      <body>
-        {/* ★ 모든 페이지 공통 래퍼: 700px 캡 + 중앙정렬 */}
-        <main className="page-wrap">
-          {children}
-        </main>
-      </body>
-    </html>
-  );
+	return (
+		<html lang='ko'>
+			<body className='min-w-[375px] max-w-[700px] mx-auto w-full bg-[#ffffff]'>
+				<Providers>
+					<main>{children}</main>
+				</Providers>
+			</body>
+		</html>
+	);
 }
