+++ src/app/planning/airline/modals/FlightDetailModal.js
