+++ src/app/planning/airline/payment/components/TermsAgreement.js
