+++ src/app/planning/airline/payment/modals/BookingDetailModal.js
