+++ src/app/providers.js
@@ -1,0 +1,6 @@
+'use client';
+import { SessionProvider } from 'next-auth/react';
+
+export default function Providers({ children }) {
+	return <SessionProvider>{children}</SessionProvider>;
+}
